# Improve CDATA parse performance