--- a/test/parse/test_cdata.rb
+++ b/test/parse/test_cdata.rb
@@ -7,10 +7,28 @@ module REXMLTests
   class TestParseCData < Test::Unit::TestCase
     include Test::Unit::CoreAssertions
 
+    def parse(xml)
+      REXML::Document.new(xml)
+    end
+
     def test_linear_performance_gt
       seq = [10000, 50000, 100000, 150000, 200000]
       assert_linear_performance(seq, rehearsal: 10) do |n|
-        REXML::Document.new('<description><![CDATA[ ' + ">" * n + ' ]]></description>')
+        parse('<description><![CDATA[ ' + ">" * n + ' ]]></description>')
+      end
+    end
+
+    class TestInvalid < self
+      def test_unclosed_cdata
+        exception = assert_raise(REXML::ParseException) do
+          parse("<root><![CDATA[a]></root>")
+        end
+        assert_equal(<<~DETAIL, exception.to_s)
+          Malformed CDATA: Missing end ']]>'
+          Line: 1
+          Position: 25
+          Last 80 unconsumed characters:
+        DETAIL
       end
     end
   end