@kou 2025-03-01
Could you also add `benchmark/parse_cdata.yaml`?